+++ lib/example/main.dart
@@ -1,6 +1,0 @@
-import 'package:fapshi_pay/example/app.dart';
-import 'package:flutter/material.dart';
-
-main() {
-  runApp(const DemoApp());
-}