+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.0.2
+
+- Updated Documentation of the Fapshi Payment button
+- Updated package description
+- Added Example file with sample implementation
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+- Initial Release of sdk with the actions;
+  - Direct Payment
+  - Check Payment Status
+  - Customization of Fapshi Button